Knossos C++ parser for MLIR interoperability
This is a draft implementation of a simplified parser for Knossos IR, so
that we can instantiate the MLIR back-end and produce its core dialect.

The AST has a few nodes, the tokeniser is very simple s-expression
descent reader and the parser interprets the semantic value of each
expression into its Knossos equivalent node.

For now, only scalar types are interpreted, global variables (no scope),
only a few operations, functions (decl, def, call) and conditionals (if/else).

There are a number of tests that check for all types, values and general
structure of the AST. Those should sufice for the time being, but
they're brittle in the sense that they only use asserts (in tests and
code).

Some simplified MLIR is also generated form the AST, and that into MLIR.
Support will increase on both sides until we have a concrete function that
can be lowered into LLVM IR.

If that works, and we see benefit, we can continue addressing the
remaining issues below.

The things left to do in the Knossos AST side:
 - Add nodes: fold, lambda, tuple, apply, assert
 - Add types: lambda, vector, tuple, linear maps
 - Add all operators (comparators, dot, etc)
 - Fully understand @types, $versions semantics

Code quality issues that need addressing:
 - Move from assert to error messages (not using exceptions)
 - Implement SSA naming scheme to functions/variables names
 - Use a robust testing driver (gtest?) intead of just asserts
 - Re-use the SymbolTeble for the MLIR back-end (type checks)