Emitting LLVM IR, with test
Tested the LLVM IR with lli and it works, but running a JIT is not as
simple. It seems calling a function "main" messes up the internal state
and the pointer to the engine itself gets lost.

Since we don't really care about native execution, I'll leave that for
later and just make sure we can execute the LLVM output. This will be
the case for LIT+FileCheck tests soon to come.