Implement vectors as memref (+build, index)
For now, only supporting fully-ranked (constexpr dimension) vectors as
ranked memref of a native type. Not supporting unranked vectors (which
include passing as a function argument) or vectors of other vectors (or
tuples).

To allow testing, adding implementation of (build) and (index). Those
are rough attempts at getting it working and need to be factored out
correctly, so we can have helper functions to build and access elements
from vectors as a generic compiler API.

Testing a build+index examples in unit test and Lit.

Lowering to MLIR and LLVM. The LLVM code was executed and found to
produce the expected results.