Adding tuple, get
This require a big refactory of the code to return multiple values/types
in MLIR, so that we can serialise the tuples (and their types). This is
the recommended strategy when lowering to standard dialect, so we'll
have to do it one day, even with a specific dialect.

Testing AST for basic syntax, simple inline MLIR get-tuple and assigning
to variables in let (then using it later) works, as well as passing
arguments to functions, returning tuples from calls, etc.