Restricting block syntax accurately
Blocks cannot be indiscriminately used in Knossos. While it was useful
for initial testing, we can now use (let) to have multiple tests in the
same function (by assiging to different variables and only returning the
last one).

Also, moving all AST tests to the cpp unit test and only leaving MLIR
and LLVM on the Lit/FileCheck tests.